// src/servicios-l/dto/create-servicios-l.dto.d.ts
export declare class CreateServiciosLDto {
    nombre: string;
    cel_guardia: string;
    codigo_postal: string;
    convenio: boolean;
    direccion: string;
    email: string;
    observacion: string;
    telefono: string;
    equipot_id: number;
    usuario_alta_id: number;
    usuario_modif_id: number;
    provincia_id: number;
}